add non-NX_FATAL library
Indended for use when there is a need to call a program's functions
manually from within a debugger.

To call a debugee's function, gdb sets up a stack frame with arguments
to the function, and return address pointing to a break-into-debugger
instruction. The instruction itself is placed on the stack. Be the stack
executable or not, when the function returns, debugger would get the
control back - either via the instruction, or a SIGSEGV due to a fetch
fault on it.

But, if both the stack is non-executable, and fetch faults are fatal, -
which is true when libprctl_wx has been preloaded - then attempting to
return would lead to SIGKILL, and not an 'ptrace-permitting' SIGSEGV.